SERVER-43883: Enable WTs cursor copy mode for ASAN builds.
When a cursor changes its position, accessing the address of the
previous key and value is illegal, but often works by accident in production.

Cursor copy mode will allocate memory for a cursor's keys and values when
positioned. When the cursor changes its position, that memory is freed. This
allows address sanitizer builds to catch bugs in MongoDBs use of WiredTiger.