SERVER-48278 Fix waits for RSM removal in sharding/remove2.js test
`sharding/remove2.js` is waiting for the ReplicaSetMonitor timeout in order to be able re-add a shard after its removal.

Instead of waiting 60s for the RSM to be removed from the interesting nodes, pols the connectionsPoolStats of these nodes until it is actually removed.

The new polling approach is reducing the test execution time on Enterprise RHEL 6.2  run time form ~2:40 to ~1:40

This commit also unblacklist the sharding/remove2.js test from the multiversion suites. Because the new polling approach solves BF-17247.