SERVER-48597 Fix stack trace unit tests with TSAN
Stack trace unit tests were failing with TSAN due to the mechanisms and
assumptions of the tests. In one case, we simulate a real failure and do
some thread necromancy to recover from it, which was causing TSAN to
segfault immediately. In another two, we rely on a feature that makes
some assumptions about which threads are running and what signals they
can receive, which TSAN's internal threads violate. These tests and
features are disabled with TSAN for now so we can run further tests with
TSAN.

Separately, one of the tests was failing with a complaint from TSAN that
errno was being stomped on. That is also addressed.