SERVER-48443 Fix builds with Icecream 1.2+ and gcc 4.4+
A bug spotted in Icecream 1.2+ can cause build failures when building
with gcc. This is, in turn, due to a bug in GCC where the preprocessor
executed via `gcc -E` has different behavior than the one used
internally during compilation. We are working with Icecream, and GCC
to address these problems. For now, we work around the bugs.

    * GCC bug report: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=88475
    * Icecream bug report: https://github.com/icecc/icecream/issues/550