SERVER-50145 Collection instances are modified using copy-on-write
When a writable collection is requested from the catalog a cloned instance will be returned.
Its lifetime is managed by default in a WriteUnitOfWork and committed into the catalog in a commit handler.
In the case of a rollback the cloned and modified collection instance is simply discarded.

This removes the need to setup rollback handlers to restore any state written to a collection in the case of a rollback.