SERVER-56877 track uncommitted changes to multikey paths with a counter
This change applies the technique in commit 9ec665b255df0e73b804f0056e48bca0dfaf7c0f
for the isMultikey boolean flags to the reader and writer multikey paths.

This commit is designed to support pre-5.0 backports and will be superseded by a
new patch that uses the Collection metadata introduced in 5.0.