SERVER-57798 Direct user to manually install on Compass download failure
We really don't need the user to tell us something has failed in a way
we generally expected might happen from time to time. It's much more
helpful from the user's perspective to know how to install Compass
manually when the installer script fails, so we'll direct them to the
appropriate documentation from now on.