SERVER-58285 Disallow MONGO_VERSION=0.0.0
Developers have commonly set MONGO_VERSION=0.0.0 in the past to avoid
having to set one explicitly for every build. However, there are better
and widely-available ways to set this automatically now, and in upcoming
releases, it will be important for MONGO_VERSION to be set to an
accurate value. From now on, we will specifically disallow setting it to
0.0.0 as a way of signaling to developers that they should stop setting
it explicitly. In the future, we may do further checking to ensure the
variable is set to an accurate value.