SERVER-68516 Revert back to always using `ProjectionExecutor` to translate projections to ABT
In SERVER-66846, we started using `ProjectionAST` as the input to ABT translations for inclusion/exclusion projections. SERVER-68516 was originally intended to update `$addFields` and `$replaceRoot` projections to also use the `ProjectionAST` codepath. However, we realized that `ProjectionAST` is not capable of representing these projections. So rather than losing functionality or maintaining two translation codepaths, we revert back to using `ProjectionExecutor` for translation.

Note this patch is not a clean revert of SERVER-66846. The diffs are:
* Keeping the tests which the original commit introduced
* Adding additional fallback logic for agg expressions and ensuring that the `_rootReplacementExpression` of a `ProjectionExecutor` is checked during the fallback mechanism. This was discovered because of a new test-case, added after  SERVER-66846, for falling back to classic when a find query has a positional projection.
* Updated comments