SERVER-76673 Reduce size of BSONColumn iterators
The iterator size is roughly cut to a third by using variant to separate
state needed for various decoding modes and states.

Also added interfaces to use the iterators as non-STL iterators, they
were already aware when they reached the end so no need for users to also
carry around an end iterator to compare against.