SERVER-77606 Fix issue where reads fail to find collection.
Reading concurrently with a DDL operation adding an index to the index catalog and also doing some multikey changes can fail to find the collection.

We reach a fallback in establishConsistentCollection where the metadata of either instance doesn't match the snapshot. When instantiating the collection for the read we need to use the collection in pending commit as the base as it might contain an index visible in the storage engine snapshot.