SERVER-76839 Remove committed flag in Collection
This fixes inconsistent visibility where readers all of a sudden discovered collections to be visible even though they had a stashed snapshot of the catalog.

It also allows for a lot of simplifications not having to deal with "waiting for visibility" and collections are now fully visible when onCommit handlers execute.