SERVER-77895 Fix window of inconsistency with collection drop
Make sure that commit handlers that add idents for pending drop to the reaper execute before drops are made visible in the catalog. If this executes in the wrong order a reader could fail to find the collection in both the reaper and catalog.

This is intended as a temporary fix until we have a long-term solution for our atomicity issues with commit handlers