SERVER-78398 Retry createIndex commands which now get aborted from setFCV command
After SERVER-77387, an ongoing createIndex command can be aborted with error 'IndexBuildAborted' if a setFeatureCompatibilityVersion command is run at the same time. This only occurs in suite
fcv_upgrade_downgrade_replica_sets_jscore_passthrough which sends setFCV command continously.
In that suite, we should retry any createIndex command that failed with that error.