SERVER-78504 Allow expressions to be children of shardFilter function call
When we initially implemented lowering for `FunctionCall['shardFilter]`,
we made an incorrect assumption that all children of that ABT will
always be of type Variable. While physical rewrites might currently only
produce such plans, path fusion rewrites can inline the definition of
variables with expressions. As such, we need to support lowering such
plans.

We do this by invoking `SBEExpressionLowering` on any child of the
shardFilter function call.