SERVER-79991 Fallback to classic when we encounter a collection with a hashed shard key.
SERVER-78253 changed the behavior of dropIndexes() to drop hashed
indexes which are backing hashed shard keys. The fallback mechanism was
assuming that a hashed shard key will always have a backing hashed
index; now this assumption is no longer valid. This change explicitly
falls back when encountering a hashed shard key to account for cases
where the backing index is dropped.