SERVER-80202 Remove optimizer test reliance on specific projection name
There are situations where we introduce several alternatives with the
same priority into the queue. Depending on the order in which these
rewrites are attempted, we might generate different projection names,
which breaks golden testing on the plan output.

This patch removes such assertions in two tests and replaces them with
broader assertions about the structure of the resulting plan.