SERVER-80255 Refactor lookup local read test to be more robust
`lookup_unionWith_subpipeline_local_read.js` is currently structured to
execute aggregations with subpipelines in various toplogoies/options and
query the profiler to assert that a certain number of reads for the
subpipeline occured as "local reads" vs "remote reads". Due to many race
conditions, the exact number might vary from run to run. This patch
refactors the test to instead assert on whether local/remote reads were
done on a particular shard.