SERVER-78802 Reject concurrent setQuerySettings on replica sets
Until now, a new value of a cluster parameter was just replacing the old one.
For the query settings project it is important to read / modify / write
cluster parameter values. In this scenario, concurrent calls of
'setQuerySettings' command may lead to seemingly successful, but incorrect
results.
This change introduces an additional 'previousTime' parameter to ensure
that the cluster parameter value that was read has not changed at the time
of writing, otherwise the operation must fail.