SERVER-80522 [CQF] Handle missing shard keys when shard filtering
Components of the shard key can be omitted from user documents. In this
case in the classic engine, the value for those component is clamped to an
explicit `null` when constructing a `ShardKeyPattern`. This is necessary because
the `ShardFilterer` implementation expects all components of the shard key to
have an explicit value. This patch implements this behavior for SBE plans
generated by Bonsai.