SERVER-80597 Ensure query settings are present on mongos before removing them in query_settings_cmds_validation.js
Due to timing issue it may happen that `removeAllQuerySettings()` doesn't remove
"all" query settings, because at the moment of deletion, not all of them are
present on mongos.

This patch resolves this issue, by asserting that the newly inserted query
settings are present on mongos as well as ensuring that the query settings are
no longer present upon deletion.