SERVER-77469 Propagate QuerySettings from mongos to the shards
Before this change, query settings were only looked up on mongod, which means that query settings will not be applied to find queries that were modified on mongos. In order to fix this issue, query settings will be looked up on mongos in sharded cluster and on mongod in replica set deployments. Upon the lookup, query settings will be propagated to the shards as command's new 'querySettings' attribute (marked as unstable).

In addition, as a part of this change, the encoding of a CanonicalQuery into a QueryShapeString will omit the 'hint' field if it is made redundant by the matching query setting.