SERVER-83461 Crash->throw on requiredToRunOnMongos/canRunOnMongos conflict
Since SERVER-65534 pipelines starting with the '$queue' stage must run on
mongoS when no 'CollectionRoutingInfo' is available. This, however, can
result in an invariant failure, because some pipelines starting with the
'$queue' stage may still not be allowed to run on mongoS.

In this fix we replace the invariant with a 'uassert', which was already
there, but did not cover the case introduced in SERVER-65534.