SERVER-83422 Remove explain from AggCmdShape
Previously 'explain' field was part of the AggCmdShape. This was causing complications in situations when users specify query settings on an aggregate command and then issue an explain to see if query settings are indeed used. Since 'explain' was part of the shape, no query settings would be reported in explain.